HW-6.2. дополнил удалением user
