+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/CategoryExtension.java
@@ -1,12 +1,11 @@
 package guru.qa.niffler.jupiter.extension;
 
 import static utils.RandomDataUtils.randomCategoryName;
-
 import guru.qa.niffler.jupiter.annotation.Category;
 import guru.qa.niffler.jupiter.annotation.meta.User;
 import guru.qa.niffler.model.spend.CategoryJson;
 import guru.qa.niffler.service.SpendClient;
-import guru.qa.niffler.service.SpendDbClient;
+import guru.qa.niffler.service.SpendJsonClient;
 import java.util.Optional;
 import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
@@ -23,7 +22,7 @@ public class CategoryExtension implements
 
   public static final ExtensionContext.Namespace NAMESPACE = ExtensionContext.Namespace.create(
       CategoryExtension.class);
-  private final SpendClient spendClient = new SpendDbClient();
+  private final SpendClient spendClient = new SpendJsonClient();
 
   @Override
   public void beforeEach(ExtensionContext context) {
