+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/CategoryTest.java
@@ -3,15 +3,12 @@ package guru.qa.niffler.test.web;
 import com.codeborne.selenide.Selenide;
 import guru.qa.niffler.config.Config;
 import guru.qa.niffler.data.entity.spend.CategoryEntity;
-import guru.qa.niffler.data.entity.spend.SpendEntity;
 import guru.qa.niffler.jupiter.annotation.Category;
 import guru.qa.niffler.jupiter.annotation.meta.User;
 import guru.qa.niffler.jupiter.extension.BrowserExtension;
 import guru.qa.niffler.model.spend.CategoryJson;
-import guru.qa.niffler.model.spend.CurrencyValues;
 import guru.qa.niffler.page.LoginPage;
-import guru.qa.niffler.service.SpendDnEntityClient;
-import java.util.Date;
+import guru.qa.niffler.service.SpendDbClient;
 import java.util.Optional;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
@@ -66,42 +63,42 @@ public class CategoryTest {
 
   @Test
   void findCategoryById() {
-    SpendDnEntityClient spendDnEntityClient = new SpendDnEntityClient();
-    Optional<CategoryEntity> categoryById = spendDnEntityClient.findCategoryById(
+    SpendDbClient spendDbClient = new SpendDbClient();
+    Optional<CategoryEntity> categoryById = spendDbClient.findCategoryById(
         UUID.fromString("0a091f06-b5bb-11f0-93fc-aa5c32f82d84"));
     System.out.println(categoryById);
   }
 
   @Test
   void findCategoryByUsernameAndSpendName() {
-    SpendDnEntityClient spendDnEntityClient = new SpendDnEntityClient();
-    Optional<CategoryEntity> categoryByUsernameAndSpendName = spendDnEntityClient.findCategoryByUsernameAndSpendName(
+    SpendDbClient spendDbClient = new SpendDbClient();
+    Optional<CategoryEntity> categoryByUsernameAndSpendName = spendDbClient.findCategoryByUsernameAndSpendName(
         "duck", "Машина");
     System.out.println(categoryByUsernameAndSpendName);
   }
 
   @Test
   void createCategoryTest() {
-    SpendDnEntityClient spendDnEntityClient = new SpendDnEntityClient();
+    SpendDbClient spendDbClient = new SpendDbClient();
 
     CategoryEntity categoryEntity = new CategoryEntity();
     categoryEntity.setName("Test1qaa");
     categoryEntity.setUsername("duck");
     categoryEntity.setArchived(false);
 
-    spendDnEntityClient.createCategory(categoryEntity);
+    spendDbClient.createCategory(categoryEntity);
   }
 
   @Test
   void removeCategoryTest() {
-    SpendDnEntityClient spendDnEntityClient = new SpendDnEntityClient();
+    SpendDbClient spendDbClient = new SpendDbClient();
 
     CategoryEntity categoryEntity = new CategoryEntity();
     categoryEntity.setName("Test5qaa");
     categoryEntity.setUsername("duck");
     categoryEntity.setArchived(false);
-    spendDnEntityClient.createCategory(categoryEntity);
-    spendDnEntityClient.removeCategory(categoryEntity);
+    spendDbClient.createCategory(categoryEntity);
+    spendDbClient.removeCategory(categoryEntity);
   }
 
 }
