+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/TransactionTest.java
@@ -19,7 +19,7 @@ import guru.qa.niffler.model.spend.CategoryJson;
 import guru.qa.niffler.model.spend.CurrencyValues;
 import guru.qa.niffler.model.spend.SpendJson;
 import guru.qa.niffler.model.user.UserJson;
-import guru.qa.niffler.service.SpendDbClient;
+import guru.qa.niffler.service.SpendJsonClient;
 import guru.qa.niffler.service.UserDbClient;
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -38,8 +38,8 @@ public class TransactionTest {
 
   @Test
   void txTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    SpendJson spendJson = spendDbClient.createSpend(
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    SpendJson spendJson = spendJsonClient.createSpend(
         new SpendJson(
             null,
             new Date(),
@@ -98,14 +98,14 @@ public class TransactionTest {
 
   @Test
   public void findCategoriesByUsernameTest() {
-    SpendDbClient dbClient = new SpendDbClient();
+    SpendJsonClient dbClient = new SpendJsonClient();
     List<CategoryJson> user = dbClient.findAllCategories("duck");
     System.out.println(user);
   }
 
   @Test
   public void findSpendsByUsernameTest() {
-    SpendDbClient dbClient = new SpendDbClient();
+    SpendJsonClient dbClient = new SpendJsonClient();
     List<SpendJson> user = dbClient.findSpendsByUserName("duck");
     System.out.println(user);
   }
@@ -131,15 +131,15 @@ public class TransactionTest {
 
   @Test
   public void findCategoryAndUserTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    Optional<CategoryJson> category = spendDbClient.findCategoryByNameAndUsername("duck", "Машина");
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    Optional<CategoryJson> category = spendJsonClient.findCategoryByNameAndUsername("duck", "Машина");
     category.stream().forEach(System.out::println);
   }
 
   @Test
   public void createCategoryTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    CategoryJson categoryJson = spendDbClient.createCategory(
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    CategoryJson categoryJson = spendJsonClient.createCategory(
         new CategoryJson(
             null,
             "QAZ1qaz",
@@ -152,38 +152,38 @@ public class TransactionTest {
 
   @Test
   public void deleteCategoryTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
     CategoryEntity category = new CategoryEntity();
     category.setId(UUID.fromString("2a6a67f0-b5bb-11f0-8a0e-aa5c32f82d84"));
-    spendDbClient.deleteCategory(category);
+    spendJsonClient.deleteCategory(category);
   }
 
   @Test
   public void deleteSpendTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
     SpendEntity spend = new SpendEntity();
     spend.setId(UUID.fromString("9d8cfc1e-b5bd-11f0-bfb0-aa5c32f82d84"));
-    spendDbClient.deleteSpend(spend);
+    spendJsonClient.deleteSpend(spend);
   }
 
   @Test
   public void findAllSpendsTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    List<SpendJson> spends = spendDbClient.findAllSpends();
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    List<SpendJson> spends = spendJsonClient.findAllSpends();
     System.out.println(spends);
   }
 
   @Test
   public void findSpendByIdTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    Optional<SpendJson> spend = spendDbClient.findSpendById("1328a312-b5bc-11f0-a017-aa5c32f82d84");
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    Optional<SpendJson> spend = spendJsonClient.findSpendById("1328a312-b5bc-11f0-a017-aa5c32f82d84");
     System.out.println(spend);
   }
 
   @Test
   public void updateSpendByIdTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    SpendJson spend = spendDbClient.updateSpend(
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    SpendJson spend = spendJsonClient.updateSpend(
         new SpendJson(UUID.fromString("888ca6da-b6e3-11f0-8e67-ea06c42c5790"),
             new Date(new Date().getTime()),
             new CategoryJson(UUID.fromString("888ba94c-b6e3-11f0-8e67-ea06c42c5790"), null, null,
@@ -198,8 +198,8 @@ public class TransactionTest {
 
   @Test
   public void updateCategoryByIdTest() {
-    SpendDbClient spendDbClient = new SpendDbClient();
-    CategoryJson category = spendDbClient.updateCategory(
+    SpendJsonClient spendJsonClient = new SpendJsonClient();
+    CategoryJson category = spendJsonClient.updateCategory(
         new CategoryJson(
             UUID.fromString("9d8c393c-b5bd-11f0-bfb0-aa5c32f82d84"),
             "1q1a1a1", "duck", false)
